pyproject.toml
[project]
name = "airtable"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "pyairtable>=3.2.0",
    "python-dotenv>=1.1.1",
]
